gemm

+++ TPBLAS/include/blas_test.h
@@ -11,9 +11,9 @@
 static const double duree_cycle = ((double) 1 / (double) 2.6);
 // duree du cycle en nano seconde 10^-9
 
-void calcul_flop (char *message, unsigned int nb_operations_flottantes, unsigned long long int cycles);
+void calcul_flop (char *message, unsigned long long int nb_operations_flottantes, unsigned long long int cycles);
 
-void calcul_octet(char *message, unsigned int nb_octet, unsigned long long int cycles);
+void calcul_octet(char *message, unsigned long long int nb_octet, unsigned long long int cycles);
 
 float* new_s(unsigned int size);
